# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DentOzone clinic operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /session:
    get:
      operationId: getSession
      tags: [staff]
      summary: Resolve the signed-in clinic staff member
      responses:
        "200":
          description: Signed-in staff identity and clinic role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffSession"
        "401":
          description: Sign-in required
        "403":
          description: Account is not authorized for this clinic
...
  /staff:
    get:
      operationId: getStaff
      tags: [staff]
      summary: List clinic staff
      responses:
        "200":
          description: Staff members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: inviteStaff
      tags: [staff]
      summary: Invite a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInviteInput"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{staffId}:
    parameters:
      - in: path
        name: staffId
        required: true
        schema: { type: integer }
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Change a staff member's role or status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffUpdate"
      responses:
        "200":
          description: Updated staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
...
components:
  schemas:
...
    StaffRole:
      type: string
      enum: [owner, dentist, reception, assistant, accountant, manager]
    StaffStatus:
      type: string
      enum: [invited, active, suspended]
    StaffSession:
      type: object
      required: [id, email, name, role, status]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role: { $ref: "#/components/schemas/StaffRole" }
        status: { $ref: "#/components/schemas/StaffStatus" }
    StaffMember:
      allOf:
        - $ref: "#/components/schemas/StaffSession"
        - type: object
          required: [createdAt]
          properties:
            createdAt: { type: string, format: date-time }
    StaffInviteInput:
      type: object
      required: [email, role]
      properties:
        email: { type: string, format: email }
        name: { type: string }
        role: { $ref: "#/components/schemas/StaffRole" }
    StaffUpdate:
      type: object
      properties:
        role: { $ref: "#/components/schemas/StaffRole" }
        status: { $ref: "#/components/schemas/StaffStatus" }
        name: { type: string }
...
  securitySchemes:
    staffSession:
      type: http
      scheme: bearer
      bearerFormat: DentOzone staff session
...